Made exefs processing run natively in python. No more externally run program.
