+++ lib/exefs_toolkit/__init__.py
@@ -1,0 +1,5 @@
+from .reader import ExeFSReader
+from .builder import ExeFSBuilder
+from .structures import ExeFSFileHdr, ExeFSHdr
+
+__all__ = ['ExeFSReader', 'ExeFSBuilder', 'ExeFSFileHdr', 'ExeFSHdr']
