+++ lib/exefs_toolkit/compression/__init__.py
@@ -1,0 +1,3 @@
+from .blz import BLZ
+
+__all__ = ['BLZ']
